+++ src/main/java/domain/movement/Movable.java
@@ -1,4 +1,4 @@
-package domain.move;
+package domain.movement;
 
 import domain.position.Position;
 import java.util.Objects;
