+++ src/main/java/domain/piece/piecerole/Queen.java
@@ -1,15 +1,15 @@
 package domain.piece.piecerole;
 
-import static domain.move.Direction.E;
-import static domain.move.Direction.N;
-import static domain.move.Direction.NE;
-import static domain.move.Direction.NW;
-import static domain.move.Direction.S;
-import static domain.move.Direction.SE;
-import static domain.move.Direction.SW;
-import static domain.move.Direction.W;
+import static domain.movement.Direction.E;
+import static domain.movement.Direction.N;
+import static domain.movement.Direction.NE;
+import static domain.movement.Direction.NW;
+import static domain.movement.Direction.S;
+import static domain.movement.Direction.SE;
+import static domain.movement.Direction.SW;
+import static domain.movement.Direction.W;
 
-import domain.move.Movable;
+import domain.movement.Movable;
 import domain.position.Position;
 import java.util.List;
 import java.util.Objects;
