+++ src/test/java/domain/game/DirectionTest.java
@@ -2,7 +2,7 @@ package domain.game;
 
 import static fixture.PositionFixture.*;
 
-import domain.move.Direction;
+import domain.movement.Direction;
 import domain.position.Position;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
