+++ src/test/java/domain/position/PositionTest.java
@@ -1,6 +1,6 @@
 package domain.position;
 
-import domain.move.ChessVector;
+import domain.movement.ChessVector;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
